Add Result object
This will be used to keep track of the outputs in between every step,
and possibly from the final Stem call.

I'm using this because each step might actually return a word that is
actually in a terminal state, where it should no longer be acted upon.
If I was to just return a simple string of the word, then the outer
calling code would need to know about the exceptions for each individual
step, and I didn't like that.