Add Step1
It's step 1 of the snowball stemming algorithm, which is actually just a
bunch of operations strung together. There might actually be a better
way to structure this code to make it flow from one to the next, and I
might even do that in the future, but for now just having the different
lines with each getting a new output is fine.

It's not the prettiest code, but it works.